backend/requirements.txt
django==3.2.8
psycopg2-binary==2.9.1
djangorestframework==3.12.4